refactor: header panel
